Photo_Enhancer-main/Backend/requirements.txt
# =============================================================================
# BASE DEPENDENCIES
# =============================================================================
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-redis>=5.4.0
whitenoise>=6.6.0

# =============================================================================
# DATABASE & CACHE
# =============================================================================
psycopg[binary]>=3.1.18
redis>=5.0.0

# =============================================================================
# TASK QUEUE
# =============================================================================
celery>=5.3.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0

# =============================================================================
# CLOUD STORAGE (Optional)
# =============================================================================
boto3>=1.34.0
django-storages>=1.14.0

# =============================================================================
# IMAGE PROCESSING
# =============================================================================
Pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.0

# =============================================================================
# AI/ML (Based on your actual photo enhancement needs)
# =============================================================================
# tensorflow>=2.15.0
# torch>=2.1.0
# torchvision>=0.16.0

# =============================================================================
# SECURITY
# =============================================================================
gunicorn>=21.2.0
uvicorn[standard]>=0.27.0
python-json-logger>=2.0.0

# =============================================================================
# MONITORING & HEALTH CHECKS
# =============================================================================
# sentry-sdk>=1.40.0
# django-prometheus>=2.5.0
# django-health-check>=3.18.0

# =============================================================================
# DEVELOPMENT (uncomment for local dev)
# =============================================================================
# django-debug-toolbar>=4.3.0
# pytest>=8.0.0
# pytest-django>=4.8.0
# pytest-cov>=4.1.0
# factory-boy>=3.3.0
# freezegun>=1.4.0
